pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
bs4 = "^0.0.1"
flask = "^3.0.0"
requests = "^2.31.0"
psycopg2 = "^2.9.9"
consultor = "^0.2.0"
youtube-data-api = "^0.0.21"
ai-tools = "^0.3.9"
openai = "^1.3.7"
git-python = "^1.0.3"
werkzeug = "^3.0.1"
flask-cors = "^4.0.0"
firebase-admin = "^6.3.0"
python-dateutil = "^2.8.2"
isodate = "^0.6.1"
xds-protos = "^1.60.0"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.110.0"
flask-session = "^0.5.0"
pinecone-client = "^2.2.4"
langchain = "^0.0.352"

[tool.poetry.overrides]
"SQLAlchemy" = { version = ">=1.4,<3" }

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"